Remove test-only thread export wrapper
